Rewritten TLM integration as proper 'operator'
We define the operator launcher containing all
the information (the same for the direct and
adjoint operator); parameters, reference
trajectory, tInt and t0.

The operator itself can then be applied (as
direct or adjoint) on different perturbations
through integrate() and adjoint() method (which
now only take the perturbation as argument.