Added specShift to spectralLib.
<TODO>
It would be good to make spectralLib the base lib instead of
PeriodicGrid since the grid should be more high level.
Doing so would makes it possible to define spectral methods in
PeriodicGrid instances.
I'm not sure yet, to which degree this would call for important
modification in pyKdV and pyDVar.